requirements.txt
torch
torchvision
numpy>=2.3.0
pillow
matplotlib
facenet-pytorch
tqdm